refactor: query와 fetch를 각각 파일로 분리
